% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm10.R
\docType{data}
\name{pm10}
\alias{pm10}
\title{Tibble con i dati di pm10 per 27 stazioni della regione FRIULIVENEZIAGIULIA}
\format{
Un tibble con 8 colonne e 15627 osservazioni
}
\usage{
pm10
}
\description{
Tibble con i dati di pm10 per 27 stazioni della regione FRIULIVENEZIAGIULIA
}
\keyword{datasets}
